requirements.txt
# This is the requirements.txt file
python-dotenv
sendgrid
pandas
